+++ src/services/RevertedBlockResolver.ts
@@ -4,11 +4,14 @@ import { Logger } from 'winston';
 import BlockChainData from '../models/BlockChainData';
 import { DeleteWriteOpResultObject } from 'mongodb';
 import Settings from '../types/Settings';
+import { MappingRepository } from '../repositories/MappingRepository';
+import { LAST_BLOCK_CHECKED_FOR_MINT_EVENT } from '../constants/mappings';
 
 @injectable()
 class RevertedBlockResolver {
   @inject('Logger') logger!: Logger;
   @inject('Settings') settings!: Settings;
+  @inject(MappingRepository) mappingRepository: MappingRepository;
 
   async apply(lastSubmittedBlockId: number, nextBlockId: number, chainId: string): Promise<number> {
     if (lastSubmittedBlockId <= nextBlockId) {
@@ -24,11 +27,25 @@ class RevertedBlockResolver {
 
     this.logger.warn(`[${chainId}] Block reverted: from ${lastSubmittedBlockId} --> ${nextBlockId}`);
 
-    const blockRes = await this.revertBlockChainDatas(nextBlockId, chainId);
+    const [blockRes] = await Promise.all([
+      this.revertBlockChainDatas(nextBlockId, chainId),
+      this.revertCachedBlock(chainId, nextBlockId),
+    ]);
 
     this.logger.info(`[${chainId}] because of reverts we deleted ${blockRes.deletedCount} blocks >= ${nextBlockId}`);
 
     return blockRes.deletedCount;
+  }
+
+  private async revertCachedBlock(chainId: string, nextBlockId: number): Promise<void> {
+    const lastBlockKey = LAST_BLOCK_CHECKED_FOR_MINT_EVENT(chainId);
+    const lastCheckedBlock = await this.mappingRepository.get(lastBlockKey);
+    const lastCheckedBlockInt = lastCheckedBlock ? parseInt(lastCheckedBlock, 10) : 0;
+
+    if (lastCheckedBlockInt > nextBlockId) {
+      this.logger.warn(`[${chainId}] reverted ${lastBlockKey} to ${nextBlockId - 1}`);
+      await this.mappingRepository.set(lastBlockKey, (nextBlockId - 1).toString(10));
+    }
   }
 
   private async revertBlocks(nextBlockId: number): Promise<void> {
@@ -37,14 +54,16 @@ class RevertedBlockResolver {
   }
 
   private async revertBlockChainDatas(nextBlockId: number, chainId: string): Promise<DeleteWriteOpResultObject> {
-    this.logger.warn(`[${chainId}] deleting many BlockChainData: blockId gte ${nextBlockId}`);
-
     const deleteBlockChainData = await BlockChainData.collection.deleteMany({
       chainId: chainId,
       blockId: { $gte: nextBlockId },
     });
 
     const remainingBlockChainDatas = await BlockChainData.collection.countDocuments({ blockId: { $gte: nextBlockId } });
+
+    this.logger.warn(
+      `[${chainId}] deleting many BlockChainData: blockId gte ${nextBlockId}, remaining blocks: ${remainingBlockChainDatas}`
+    );
 
     if (remainingBlockChainDatas === 0) {
       await this.revertBlocks(nextBlockId);